requirements.txt
numpy
matplotlib
scipy
kagglehub
Pillow         
scikit-image   
POT            
pybind11

